Fixup system files
There was a bug introduced last commit for And/Or IR. On top of that,
there was a lot of duplication for reference_table, though it was
never used.

Moved `Reference` to an IR file / module to be reused. Removed
ReferenceTable and the original `Reference` in semantics. Most of
semantics will be deleted over time, probably subsumed by IR.
