+++ src/ir/mod.rs
@@ -1,4 +1,5 @@
 pub mod conditions;
+pub mod reference;
 
 trait Instruction {
     fn synthesize() -> String;
